refactor(cognition): settle_step + drive_to_settle + now_ms extracted into settle.rs (pure code-motion)
Moved the two settlement drivers (both stay pub, re-exported via
`pub use settle::{drive_to_settle, settle_step}`) plus the now_ms helper (now pub(super))
verbatim into a new settle.rs. Cross-module calls resolve via `use super::apply::apply_act`,
`use super::perception::{...}`, and `use super::types::{...}`; two relocation-forced paths
rewritten super::→crate:: (deliberation_budget, and apply.rs's now_ms import now
`super::settle::now_ms`). mod.rs top collapsed to a thin coordinator (module wiring + pub
re-exports only); the now test-only type/helper imports moved down into the tests mod's own
`use` block, leaving zero unused-import warnings in act_observe. Behavior-identical; no external
caller changes. mod.rs 2159 → 1600 lines; settle.rs is 596 lines. act_observe tests green:
28 passed, 0 failed.
